// src/enums/infoMessages.ts
export enum TenantInfoMessages {
    TenantMiddlewareError = "Tenant middleware error",
    NewConnectionEstablished = "New connection established for tenant",
    ConnectionReused = "Connection reused for tenant",
    CreatingDatabase = "Creating Database",
    DatabaseCreatedSuccessfully = "Database created successfully",
    DatabaseCreationFailed = "Database creation failed",
    RunningMigrations = "Running Migrations",
    MigrationsCompleted = "Migrations Completed",
    MigrationsFailed = "Migrations Failed",
    DatabaseInitializedSuccessfully = "Database Initialized Successfully",
    DatabaseInitializationFailed = "Database Initialization Failed",
    
}